+++ core/src/main/scala/core/DataFrameFromTo.scala
@@ -23,6 +23,8 @@ import java.text.SimpleDateFormat
 import java.time.Instant
 import java.util
 import java.util.{Calendar, Properties, UUID}
+import javax.mail.internet.{InternetAddress, MimeMessage}
+import javax.mail.{Message, Session, Transport}
 
 import com.amazonaws.services.logs.model.{DescribeLogStreamsRequest, InputLogEvent, PutLogEventsRequest}
 import com.amazonaws.services.s3.model._
@@ -39,8 +41,6 @@ import com.mongodb.{MongoClient, MongoClientURI}
 import config.AppConfig
 import core.DataPull.jsonObjectPropertiesToMap
 import helper._
-import javax.mail.internet.{InternetAddress, MimeMessage}
-import javax.mail.{Message, Session, Transport}
 import org.apache.avro.Schema
 import org.apache.avro.generic.{GenericData, GenericRecord}
 import org.apache.hadoop.conf.Configuration
@@ -49,6 +49,7 @@ import org.apache.kafka.clients.consumer.KafkaConsumer
 import org.apache.kafka.clients.producer.{KafkaProducer, ProducerConfig, ProducerRecord}
 import org.apache.spark.SparkConf
 import org.apache.spark.rdd.RDD
+import org.apache.spark.sql.jdbc.JdbcDialects
 import org.apache.spark.sql.{Row, SaveMode, SparkSession}
 import org.bson.Document
 import org.codehaus.jettison.json.JSONObject
@@ -1313,12 +1314,30 @@ class DataFrameFromTo(appConfig: AppConfig, pipeline : String) extends Serializa
     df.write.mode(savemode).options(jdbcOptions).jdbc(url, table, connectionProperties)
   }
 
-  def hiveToDataFrame(table: String, sparkSession: SparkSession): org.apache.spark.sql.DataFrame = {
+  def hiveToDataFrame(cluster: String,sparkSession: org.apache.spark.sql.SparkSession, dbtable: String, username:String,fetchsize:String): org.apache.spark.sql.DataFrame  = {
 
-    sparkSession.conf.set("hive.exec.dynamic.partition.mode", "nonstrict")
+    import org.apache.spark.sql.jdbc.JdbcDialect
 
-    val df = sparkSession.sql("Select * FROM " + table)
-    df
+    val HiveDialect = new JdbcDialect {
+      override def canHandle(url: String): Boolean = url.startsWith("jdbc:hive2") || url.contains("hive2")
+
+      override def quoteIdentifier(colName: String): String =
+
+      { s"$colName" }
+
+    }
+
+    JdbcDialects.registerDialect(HiveDialect)
+
+    val jdbcDF = sparkSession.read
+      .format("jdbc")
+      .option("url", cluster)
+      .option("dbtable", dbtable)
+      .option("username",username)
+      // .option("fetchsize", fetchsize.toInt)
+      .load()
+
+    jdbcDF
   }
 
   def dataFrameToHive(table: String, saveMode: String, df: org.apache.spark.sql.DataFrame): Unit = {