+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
+## [0.1.12] - 2020-08-05
+### Changed
+* core/pom.xml
+* core/src/main/scala/core/DataFrameFromTo.scala
+* core/src/main/scala/core/Migration.scala
 
 ## [0.1.11] - 2020-07-28